Moved SuicideSystem.cs
Moved SuicideSystem.cs from server/chat to server/suicide
